incremental push
still need to fix login but go ahead and fry whatever needs to be on ts
